decrating: Phase 1 (auth) — collapse dual impl, preserve shim fallback
Delete the stale 1212-LOC in-tree `crates/shipper/src/auth.rs` and promote
the 333-LOC shim from `auth_micro.rs` to canonical `auth.rs`. The shim was
already the production code path (CLI defaults to `micro-all`), and it is
more than a thin re-export: it layers credential-file fallback and OIDC
detection on top of the `shipper-auth` microcrate.

Preserved shim behavior:
- Delegation to `shipper_auth::resolve_token` as first-choice path.
- Legacy fallback to `\$CARGO_HOME/credentials.toml` and the older
  `credentials` file when the microcrate returns no token.
- crates.io alias handling (`crates.io`, `crates_io`, and the
  nested-table `[registries.crates.io]` TOML shape).
- `detect_auth_type` / `detect_auth_type_from_token` helpers for
  OIDC trusted-publishing diagnostics.
- All `serial_test` unit tests covering resolution order, env
  precedence, legacy credentials file, and OIDC detection.

Public API surface is a superset of the deleted in-tree module
(`resolve_token`, `detect_auth_type`, `detect_auth_type_from_token`),
so no call-site changes were required.

Feature-flag cleanup:
- Remove `micro-auth` feature from `crates/shipper/Cargo.toml`; make
  `shipper-auth` a non-optional dependency.
- Drop `micro-auth` from the `micro-all` feature list.
- Remove `micro-auth` from `crates/shipper-cli/Cargo.toml` features.
- Remove `micro-auth` from the BDD feature-set matrix in
  `.github/workflows/ci.yml` and `templates/circleci-config.yml`.
- Update `docs/architecture.md` and `docs/testing.md` to reflect that
  `shipper-auth` is now a required dependency.

Validation:
- `cargo test -p shipper` — all tests pass (incl. 8 auth + 4 facade
  integration auth tests).
- `cargo test -p shipper-cli` — all tests pass.
- `cargo build -p shipper-cli` — succeeds.
- `cargo check --workspace` — succeeds.
- `cargo clippy -p shipper --all-targets --all-features -- -D warnings`
  — clean.

Net LOC removed: ~879 (1212 deleted in-tree file minus 333 promoted shim).
